20190826asdfsadfasfd

+++ PCBVI/Forms/Production/MaterialsManageForm.cs
@@ -4,6 +4,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
 using System.Drawing;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
@@ -35,14 +36,23 @@ namespace PCBVI.Forms.Production
 
         private void UscTopMenu_UpdateButtonClicked(object sender, Controls.CommonControl.TopMenubar.UpdateButtonClickedEventArgs e)
         {
-            uscTopMenu.UpdateAll(uscList.GetUpateList());
+            List<Data.MaterialsManage> materialsManages = uscList.GetUpateList();
+
+            if (materialsManages == null)
+                return;
+            
+            uscTopMenu.UpdateAll(materialsManages);
+            MessageBox.Show("저장이 완료되었습니다.", "알림", MessageBoxButtons.OK, MessageBoxIcon.Information);
         }
 
         private void UscTopMenu_DeleteButtonClicked(object sender, Controls.CommonControl.TopMenubar.DeleteButtonClickedEventArgs e)
         {
             Data.MaterialsManage materialsManage = uscList.GetCurrentLow();
+            if (materialsManage == null)
+                return;
             uscTopMenu.DeleteAt(materialsManage);
             uscList.SetDataSource(DB.MaterialsManage.Search(_lotId,_processId,_materalStateId));
+            MessageBox.Show("삭제가 완료되었습니다.", "알림", MessageBoxButtons.OK, MessageBoxIcon.Information);
         }
 
         private void UscTopMenu_InsertButtonClicked(object sender, Controls.CommonControl.TopMenubar.InsertButtonClickedEventArgs e)
@@ -53,7 +63,21 @@ namespace PCBVI.Forms.Production
         private void UscTopMenu_ExcelButtonClicked(object sender, Controls.CommonControl.TopMenubar.ExcelButtonClickedEventArgs e)
         {
             DataGridView data = uscList.GetListView();
-            uscTopMenu.SaveExcelFile(data, "자재 투입 정보");
+
+            string fileName = null;
+            string filePath = null;
+            SaveFileDialog saveFileDialog = new SaveFileDialog();
+            saveFileDialog.Title = "저장경로 지정하세요.";
+            saveFileDialog.OverwritePrompt = true;
+            saveFileDialog.Filter = " Excel97 - 2003 통합문서(*.xls) | *.xls | Excle 통합 문서(*.xlsx)| *.xlsx ";
+            //saveFileDialog.Filter = "Excle 통합 문서(*.xlsx)| *.xlsx | Excel97 - 2003 통합문서(*.xls) | *.xls";
+
+            if (saveFileDialog.ShowDialog() == DialogResult.OK)
+            {
+                fileName = saveFileDialog.FileName;
+                filePath = Path.GetFullPath(saveFileDialog.FileName);
+                uscTopMenu.SaveExcelFile(data, filePath, fileName);
+            }
         }
     }
 }
